kj;k

+++ data_decoding.h
@@ -28,8 +28,8 @@ typedef union{
 } Timestamp;
 
 typedef struct {
-		uint8_t tv_sec;
-		uint8_t tv_usec;
+		uint64_t tv_sec;
+		uint64_t tv_usec;
 } Timeval16; //redefine a new 16byte timeval for beaglebone because beaglebone has 8 byte timeval
 
 typedef union{
